Search datasets and distributions in parallel.
There are multiple ways this can be done.

In the resources config, you could simply make a superclass of both the
dataset and the distribution, and add the corresponding model to search
for. This adds some boilerplate, so we gave preference to the other
option:

Use the `[:or:]` construct from the resources API to filter on both the
datasets and their connected distributions. Ideally, the query parameter
object passed to Ember-Data's JSON-API should be translated to a form
that matches the resources API, but this was a challenge because of the
behaviour of multiple filters.

According to the documentation, `{filter: ['x', 'y']}` should turn into
`?filter=x&filter=y`, but it actually turns into
`?filter[]=x&filter[]=y`, which is not accepted by the resources API.
According to the documentation for [Store.query], it is possible to pass
`adapterOptions`, which are passed down to the adapter. Futhermore, in
the JSON-API adapter, [buildQueryParams] suggests that the option
`{arrayFormat: 'repeat'}` does what we want. However, it is unclear
where that option is picked up (if anywhere).

Therefore, I have chosen to pass the key in its entirety:
`{'filter[:or:]': …, 'filter[:or:][distributions]': …}`

Store.query: https://api.emberjs.com/ember-data/5.5/classes/Store#query
buildQueryParams: https://api.emberjs.com/ember-data/5.5/functions/@ember-data%2Fjson-api%2Frequest/buildQueryParams